Consolidate 15 ECS marker tags into app/tags/tags.h (closes #1201)
Per the ECS folder allowlist (.squad/decisions.md), app/tags/ holds a single
header — tags.h — that declares every marker tag in the codebase. Previously
the 15 tags were scattered across 6 files in components/ and entities/.

Pure relocation: tag types and their usages are unchanged. Each former host
file now `#include "tags/tags.h"` so existing consumers continue to compile
without modification.

Tags moved:
  PlayerTag, ObstacleTag, OnsetMarkerTag, ResolvedObstacleTag, ScoredTag,
  NonScorableTag, MissTag, ParticleTag, EnergyBarTag, TestPlayerPlannedTag,
  TagWorldPass, TagEffectsPass, TagHUDPass, BeatMapTag, SettingsTag.

Verification:
  - Zero `struct \w*Tag\w*\s*\{\s*\}` declarations remain outside app/tags/.
  - Build is warning-free under Clang -Wall -Wextra -Werror (native + WASM
    use the same flags).
  - shapeshifter_tests: 2956 assertions in 901 test cases pass.
